+++ src/flow/Retry.ts
@@ -1,8 +1,9 @@
 import { sleep } from './Sleep.js';
 import { resolveLogOption } from '../debug/Logger.js';
 import { since, toString as elapsedToString } from './Elapsed.js';
-import { throwIntegerTest } from '../Guards.js';
+import { throwIntegerTest, throwNumberTest } from '../Guards.js';
 import { getErrorMessage } from '../debug/GetErrorMessage.js';
+import type { Result } from '../Results.js';
 /**
  * Result of backoff
  */
@@ -33,60 +34,132 @@ export type RetryResult<V> = {
   readonly value: V | undefined;
 };
 
-/**
- * Backoff options
- */
-export type RetryOpts<V> = {
+export type BackoffOptions = {
   /**
-   * Maximum number of attempts to make
+   * Initial value.
+   * Default: 1
    */
-  readonly count: number;
+  startAt: number,
   /**
-   * Starting milliseconds for sleeping after failure
-   * Defaults to 1000.
-   * Must be above zero.
+   * Maximum times to run.
+   * Default: continues forever
    */
-  readonly startMs?: number;
+  limitAttempts: number,
   /**
-   * Initial waiting period before first attempt (optional)
+   * Stop retrying if this maximum is reached
+   * Default: no limit
    */
-  readonly predelayMs?: number;
+  limitValue: number
   /**
-   * Optional abort signal
+   * Math power. 
+   * Default: 1.1
    */
-  readonly abort?: AbortSignal;
-  /**
-   * Log: _true_ logs to console
-   */
-  readonly log?: boolean;
-  /**
-   * Math.pow factor. Defaults to 1.1. How much slower to
-   * get with each retry.
-   */
-  readonly power?: number;
-
-  /***
-   * Default value to return if it fails
-   */
-  readonly defaultValue?: V;
+  power: number
 };
 
 /**
- * Keeps calling `cb` until it returns something other than _undefined_. If it throws an exception,
- * it will cancel the retry, bubbling the exception and cancelling the retry.
+ * Generates an expoential backoff series of values
+ * ```js
+ * // Default: start at 1, power 1.1
+ * for (const v of backoffGenerator()) {
+ *  // v: numeric value
+ * }
+ * ```
+ * 
+ * By default the generator runs forever. Use either
+ * `limitAttempts` or `limitValue` to stop it when it produces a
+ * given quantity of values, or when the value itself reaches a threshold.
+ * 
+ * For example:
+ * ```js
+ * // `values` will have five values in it
+ * const values = [...backoffGenerator({ limitAttempts: 5 })];
+ * // Keep generating values until max is reached
+ * const values = [...backoffGenerator({ limitValue: 1000 })];
+ * ```
+ * 
+ * Options:
+ * * startAt: start value
+ * * limitAttempts: cap the number of values to generate
+ * * limitValue: cap the maximum calculated value
+ * * power: power value (default 1.1)
+ * 
+ * @param options 
+ * @returns 
+ */
+export function* backoffGenerator(options: Partial<BackoffOptions> = {}) {
+  const startAt = options.startAt ?? 1;
+  let limitAttempts = options.limitAttempts ?? Number.MAX_SAFE_INTEGER;
+  const limitValue = options.limitValue;
+  const power = options.power ?? 1.1;
+  let value = startAt;
+  throwIntegerTest(limitAttempts, `aboveZero`, `limitAttempts`);
+  throwNumberTest(startAt, ``, `startAt`);
+  throwNumberTest(limitAttempts, ``, `limitAttempts`);
+  if (limitValue !== undefined) throwNumberTest(limitValue, ``, `limitValue`);
+  throwNumberTest(power, ``, `power`);
+
+  while (limitAttempts > 0) {
+    // Value has climbed to the limit
+    if (limitValue && value >= limitValue) return;
+    limitAttempts--;
+    yield value;
+
+    // Increase value for next iteration
+    value += Math.pow(value, power);
+  }
+}
+
+/**
+ * Backoff options
+ */
+export type RetryOpts<T> = BackoffOptions & {
+  /**
+   * Initial waiting period before first attempt (optional)
+   */
+  readonly predelayMs: number;
+  /**
+   * Optional abort signal
+   */
+  readonly abort: AbortSignal;
+  /**
+   * Log: _true_ monitors the task execution by logging to console
+   */
+  readonly log: boolean;
+
+  /***
+   * Default task value to return if it fails
+   */
+  readonly taskValueFallback: T;
+};
+
+export type RetryTask<T> = {
+  /**
+   * If `probe` returns {success:true} task is considered
+   * complete and retrying stops
+   * @returns 
+   */
+  probe: (attempts: number) => Promise<Result<T>>
+}
+
+/**
+ * Keeps calling `callback` until it returns something other than _undefined_. 
+ * There is an exponentially-increasing delay between each retry attempt.
+ * 
+ * If `callback` throws an exception, the retry is cancelled, bubbling the exception.
  *
  * ```js
  * // A function that only works some of the time
  * const flakyFn = async () => {
  *  // do the thing
  *  if (Math.random() > 0.9) return true; // success
- *  return; // 'failed'
+ *  return; // fake failure
  * };
  *
  * // Retry it up to five times,
  * // starting with 1000ms interval
- * const result = await retry(flakyFn, {
- *  count: 5
+ * const result = await retryFunction(flakyFn, {
+ *  limitAttempts: 5
  * });
  *
  * if (result.success) {
@@ -100,7 +173,7 @@ export type RetryOpts<V> = {
  * An `AbortSignal` can be used to cancel process.
  * ```js
  * const abort = new AbortController();
- * const result = await retry(cb, { startMs: 6000, count: 1000, signal: abort.signal });
+ * const result = await retryFunction(cb, { signal: abort.signal });
  *
  * // Somewhere else...
  * abort('Cancel!'); // Trigger abort
@@ -109,48 +182,62 @@ export type RetryOpts<V> = {
  * @param opts Options
  * @returns
  */
-export const retry = async <V>(
-  callback: () => Promise<V | undefined>,
-  //eslint-disable-next-line functional/prefer-immutable-types
-  opts: RetryOpts<V>
+export const retryFunction = <T>(callback: () => Promise<T | undefined>, opts: Partial<RetryOpts<T>> = {}) => {
+  const task: RetryTask<T> = {
+    async probe() {
+      const v = await callback();
+      if (v === undefined) return { value: opts.taskValueFallback, success: false };
+      return { value: v, success: true };
+    },
+  }
+  return retryTask(task, opts);
+}
+
+export const retryTask = async <V>(
+  task: RetryTask<V>,
+  opts: Partial<RetryOpts<V>> = {}
 ): Promise<RetryResult<V>> => {
   const signal = opts.abort;
   const log = resolveLogOption(opts.log);
-  const power = opts.power ?? 1.1;
   const predelayMs = opts.predelayMs ?? 0;
   const startedAt = since();
 
-  //eslint-disable-next-line functional/no-let
-  let t = opts.startMs ?? 1000;
-  const count = opts.count;
-  //eslint-disable-next-line functional/no-let
   let attempts = 0;
+  const initialValue = opts.startAt ?? 1000;
+  const limitAttempts = opts.limitAttempts ?? Number.MAX_SAFE_INTEGER;
+  const backoffGen = backoffGenerator({ ...opts, startAt: initialValue, limitAttempts });
 
-  throwIntegerTest(count, `aboveZero`, `count`);
-  if (t <= 0) throw new Error(`startMs must be above zero`);
+  if (initialValue <= 0) throw new Error(`Param 'initialValue' must be above zero`);
 
-  if (predelayMs > 0) await sleep({ millis: predelayMs, signal: signal });
-  if (signal?.aborted) {
-    return {
-      success: false,
-      attempts,
-      value: opts.defaultValue,
-      elapsed: startedAt(),
-      message: `Aborted during predelay`,
-    };
+  if (predelayMs > 0) {
+    try {
+      await sleep({ millis: predelayMs, signal: signal });
+    } catch (error) {
+      // Could happen due to abort signal
+      return {
+        success: false,
+        attempts,
+        value: opts.taskValueFallback,
+        elapsed: startedAt(),
+        message: getErrorMessage(error),
+      };
+    }
   }
-  while (attempts < count) {
+
+  for (const t of backoffGen) {
     attempts++;
-    const callbackResult = await callback();
-    if (callbackResult !== undefined) {
-      return { value: callbackResult, success: true, attempts, elapsed: startedAt() };
+
+    // Run task
+    const result = await task.probe(attempts);
+    if (result.success) {
+      return { success: result.success, value: result.value, attempts, elapsed: startedAt() };
     }
     log({
       msg: `retry attempts: ${ attempts } t: ${ elapsedToString(t) }`,
     });
 
     // Did not succeed.
-    if (attempts >= count) {
+    if (attempts >= limitAttempts) {
       break; // Out of attempts, no point sleeping again
     }
     // Sleep
@@ -161,21 +248,18 @@ export const retry = async <V>(
       return {
         success: false,
         attempts,
-        value: opts.defaultValue,
+        value: opts.taskValueFallback,
         message: getErrorMessage(error),
         elapsed: startedAt(),
       };
     }
-
-    // Increase sleep time for next fail
-    t = Math.floor(Math.pow(t, power));
   }
 
   return {
     message: `Giving up after ${ attempts } attempts.`,
     success: false,
     attempts,
-    value: opts.defaultValue,
+    value: opts.taskValueFallback,
     elapsed: startedAt(),
   };
 };